@@ -25,12 +25,12 @@ ext {
 }
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 31
     buildToolsVersion "27.0.3"
 
     defaultConfig {
         minSdkVersion 17
-        targetSdkVersion 27
+        targetSdkVersion 31
         versionCode 4
         versionName "1.3.1"
     }
@@ -43,10 +43,11 @@ android {
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.android.support:appcompat-v7:27.1.1'
-    implementation 'com.android.support:support-annotations:27.1.1'
-    implementation 'com.android.support:support-v4:27.1.1'
+    implementation 'androidx.appcompat:appcompat:1.4.0'
+    implementation 'androidx.annotation:annotation:1.3.0'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     implementation 'com.github.dcendents:android-maven-gradle-plugin:2.0'
+    implementation 'androidx.fragment:fragment:1.4.0'
 }
 // build a jar with source files
 task sourcesJar(type: Jar) {
